MMC3 ethernet example: Added simulation
